+++ src/parse.js
@@ -1,4 +1,4 @@
-function parsePaths(paths) {
+function parseEntryPaths(paths) {
   const ret = paths || [];
 
   // Convert possible string to an array
@@ -9,6 +9,21 @@ function parsePaths(paths) {
   return ret;
 }
 
+function parseEntries(paths, chunkName) {
+  if (Array.isArray(paths)) {
+    return paths;
+  }
+
+  if (!(chunkName in paths)) {
+    throw new Error(`Failed to find ${chunkName} in ${paths.join(', ')}`);
+  }
+
+  const ret = paths[chunkName];
+
+  return Array.isArray(ret) ? ret : [ret];
+}
+
 module.exports = {
-  paths: parsePaths
+  entryPaths: parseEntryPaths,
+  entries: parseEntries
 };